chore: rename package to react-router-fastify
