+++ .changeset/react-router-fastify-rewrite.md
@@ -1,5 +1,5 @@
 ---
-"@mcansh/remix-fastify": major
+"@mcansh/react-router-fastify": major
 ---
 
 Rebuild the package as a React Router 8 framework-mode Fastify adapter.
@@ -7,7 +7,7 @@ Rebuild the package as a React Router 8 framework-mode Fastify adapter.
 This is a breaking rewrite of the public API and package shape. The Remix-focused adapter APIs have been removed:
 
 - Removed `remixFastify` and Remix request-handler support.
-- Removed the `@mcansh/remix-fastify/middleware` and `@mcansh/remix-fastify/react-router` subpath exports.
+- Removed the `@mcansh/react-router-fastify/middleware` and `@mcansh/react-router-fastify/react-router` subpath exports.
 - Removed the generated CommonJS proxy files for those subpaths.
 - Removed CommonJS package output. The package is now ESM-only and publishes `./dist/index.mjs` plus `./dist/vite.mjs`.
 - React Router 8, `@react-router/node` 8, Fastify 5, and Node `>=22.22.0` are now required.
