+++ .changeset/react-router-fastify-vite-dev.md
@@ -1,8 +1,8 @@
 ---
-"@mcansh/remix-fastify": major
+"@mcansh/react-router-fastify": major
 ---
 
-Add the `@mcansh/remix-fastify/vite` entry point with `fastifyReactRouterDev`.
+Add the `@mcansh/react-router-fastify/vite` entry point with `fastifyReactRouterDev`.
 
 The Vite plugin lets `react-router dev` run through the app's own Fastify server instead of a separate manual server process. It loads a configurable server entry, calls the configured factory export, waits for `app.ready()`, and mounts Fastify after Vite's own middleware so Vite internals and HMR continue to run first.
