+++ .changeset/remix-fastify-package-rename.md
@@ -1,0 +1,9 @@
+---
+"@mcansh/remix-fastify": major
+---
+
+Rename the package to `@mcansh/react-router-fastify`.
+
+`@mcansh/remix-fastify` is now a compatibility bridge that re-exports the new
+package. Update application dependencies and imports to
+`@mcansh/react-router-fastify`.
