+++ .github/workflows/release.yml
@@ -86,4 +86,4 @@ jobs:
         uses: remix-run/release-comment-action@v0.5.1
         with:
           DIRECTORY_TO_CHECK: "./packages"
-          PACKAGE_NAME: "@mcansh/remix-fastify"
+          PACKAGE_NAME: "@mcansh/react-router-fastify"
