+++ README.md
@@ -1,6 +1,6 @@
-# remix-fastify
+# react-router-fastify
 
-Workspace for `@mcansh/remix-fastify`, a Fastify adapter and Vite development plugin for React Router v8 framework mode.
+Workspace for `@mcansh/react-router-fastify`, a Fastify adapter and Vite development plugin for React Router v8 framework mode.
 
-- Package documentation: [`packages/remix-fastify`](./packages/remix-fastify)
+- Package documentation: [`packages/react-router-fastify`](./packages/react-router-fastify)
 - Example app: [`examples/basic`](./examples/basic)
