+++ examples/basic/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "remix-fastify-basic-example",
+  "name": "react-router-fastify-basic-example",
   "private": true,
   "type": "module",
   "imports": {
@@ -12,7 +12,7 @@
     "typecheck": "react-router typegen && tsc"
   },
   "dependencies": {
-    "@mcansh/remix-fastify": "workspace:*",
+    "@mcansh/react-router-fastify": "workspace:*",
     "@react-router/node": "8.0.1",
     "fastify": "^5.8.5",
     "isbot": "^5",
