+++ package.json
@@ -1,18 +1,18 @@
 {
-  "name": "remix-fastify-workspace",
+  "name": "react-router-fastify-workspace",
   "private": true,
   "license": "MIT",
   "repository": "mcansh/remix-fastify",
   "type": "module",
   "scripts": {
-    "build": "pnpm --filter @mcansh/remix-fastify run build",
+    "build": "pnpm --recursive --filter './packages/*' run build",
     "clean": "git clean -fdX",
     "example:build": "pnpm --dir examples/basic run build",
     "example:dev": "pnpm --dir examples/basic run dev",
     "example:start": "pnpm --dir examples/basic run start",
     "example:typecheck": "pnpm --dir examples/basic run typecheck",
-    "test": "pnpm --filter @mcansh/remix-fastify run test",
-    "typecheck": "pnpm --filter @mcansh/remix-fastify run typecheck",
+    "test": "pnpm --filter @mcansh/react-router-fastify run test",
+    "typecheck": "pnpm --recursive --filter './packages/*' run typecheck",
     "validate": "pnpm run typecheck && pnpm run test && pnpm run build && pnpm run example:typecheck && pnpm run example:build",
     "changeset": "changeset",
     "changeset:version": "changeset version && node ./scripts/remove-prerelease-changelogs.ts && node ./scripts/post-changeset.ts",
