+++ packages/react-router-fastify/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@mcansh/remix-fastify",
+  "name": "@mcansh/react-router-fastify",
   "version": "4.1.2",
-  "description": "Compatibility bridge for @mcansh/react-router-fastify",
+  "description": "Fastify adapter and Vite development plugin for React Router v8 framework mode",
   "keywords": [
     "adapter",
     "fastify",
@@ -16,7 +16,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/mcansh/remix-fastify.git",
-    "directory": "packages/remix-fastify"
+    "directory": "packages/react-router-fastify"
   },
   "funding": [
     {
@@ -25,7 +25,10 @@
     }
   ],
   "files": [
+    "CHANGELOG.md",
     "dist",
+    "src",
+    "!src/**/*.test.ts",
     "package.json",
     "README.md",
     "LICENSE"
@@ -45,17 +48,28 @@
     "build": "tsdown",
     "clean": "git clean -fdX",
     "prepublishOnly": "pnpm run build",
+    "test": "vitest",
     "typecheck": "tsc --noEmit",
-    "validate": "pnpm run typecheck && pnpm run build"
+    "validate": "pnpm run typecheck && pnpm run test && pnpm run build"
   },
   "dependencies": {
-    "@mcansh/react-router-fastify": "workspace:^"
+    "@fastify/static": "^9.1.3"
   },
   "devDependencies": {
     "@arethetypeswrong/core": "^0.18.3",
+    "@react-router/dev": "8.0.1",
+    "@react-router/node": "8.0.1",
+    "@types/node": "^24.10.1",
+    "@vitest/coverage-v8": "4.1.9",
+    "fastify": "^5.8.5",
     "publint": "^0.3.21",
+    "react": "^19.2.7",
+    "react-dom": "^19.2.7",
+    "react-router": "8.0.1",
     "tsdown": "^0.22.3",
-    "typescript": "^5.9.3"
+    "typescript": "^5.9.3",
+    "vite": "^8.0.16",
+    "vitest": "^4.0.15"
   },
   "peerDependencies": {
     "@react-router/node": "^8.0.0",
