+++ packages/react-router-fastify/src/handler.ts
