+++ packages/react-router-fastify/src/response.ts
