+++ packages/react-router-fastify/src/vite-runtime.ts
