+++ packages/remix-fastify/src/index.ts
@@ -1,13 +1,1 @@
-export { fastifyReactRouter } from "./fastify.ts"
-export type { FastifyReactRouterOptions } from "./fastify.ts"
-export { createRequestHandler } from "./handler.ts"
-export { createHeaders, createRequest, createUrl } from "./request.ts"
-export { sendResponse } from "./response.ts"
-export type {
-  CreateRequestHandlerOptions,
-  GetLoadContextFunction,
-  HttpRequest,
-  HttpResponse,
-  ReactRouterLoadContext,
-  RequestHandler,
-} from "./handler.ts"
+export * from "@mcansh/react-router-fastify"
